+++ src/pages/Lol.jsx
@@ -91,7 +91,7 @@ export default function Lol() {
           />
         ))}
 
-        <div className="flex justify-center mt-4 bg-transparent">
+        <div className="mt-4 flex justify-center bg-transparent">
           <button
             className="mt-[20px] flex w-[300px] items-center justify-center rounded-md border border-transparent bg-gradient-to-r from-purple-600 to-indigo-600 px-4 py-2 text-[30px] text-white shadow-sm hover:from-purple-700 hover:to-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
             type="submit"
