+++ src/util/tierPoints.js
@@ -1,0 +1,11 @@
+export const tierToPoints = {
+  Iron: 7,
+  Bronze: 13,
+  Silver: 17,
+  Gold: 25,
+  Platinum: 29,
+  Diamond: 37,
+  Ascendant: 43,
+  Immortal: 45,
+  Radiant: 48,
+};
